+++ Properties/AssemblyInfo.cs
@@ -5,11 +5,36 @@ using System.Runtime.InteropServices;
 // General Information about an assembly is controlled through the following
 // set of attributes. Change these attribute values to modify the information
 // associated with an assembly.
-[assembly: AssemblyTitle("Extensions")]
-[assembly: AssemblyDescription("Extensions.dll contains extension methods that enhance existing C# classes thus making life easier for developers.")]
+//[assembly: AssemblyTitle("Extensions")]
+[assembly: AssemblyTitle("Extensions.dll contains extension methods that enhance existing C# classes thus making life easier for developers." +
+                 "\nExtensions.dll supports all versions of.NET from 3.5 through 5.0" +
+                 "\nHighlights are methods such as:" +
+                 "\nUse.Save() and.Load() for easy state saving of any object in C#." +
+                 "\nUse.Retry() to easily handle 429 throttling errors in web calls." +
+                 "\nUse.IsPrime() to determine if a number is a prime number." +
+                 "\nUse.GetNthPrime() and.GetNthPrimeAsync to get the Nth prime number capable of calculating 10, 000, 000 primes / minute on an Intel Core i7 - 6700K CPU @ 4.00 GHz using 7 threads on 8 logical cores." +
+                 "\nUse.ReplaceTokens() to replace multiple strings in string dictionary style." +
+                 "\nUse.IsStrong() to validate strong passwords." +
+                 "\nUse.ToTimeZone() to quickly convert any given time to any given time zone." +
+                 "\nUse.Elevate() to restart current process in admin mode if it doesn't have admin rights." +
+                 "\nUse.CompoundInterest() to calculate any given interest over any given perior at any given interval." +
+                 "\nand many, many more.")]
+[assembly: AssemblyDescription("Extensions.dll contains extension methods that enhance existing C# classes thus making life easier for developers." +
+                 "\nExtensions.dll supports all versions of.NET from 3.5 through 5.0" +
+                 "\nHighlights are methods such as:" +
+                 "\nUse.Save() and.Load() for easy state saving of any object in C#." +
+                 "\nUse.Retry() to easily handle 429 throttling errors in web calls." +
+                 "\nUse.IsPrime() to determine if a number is a prime number." +
+                 "\nUse.GetNthPrime() and.GetNthPrimeAsync to get the Nth prime number capable of calculating 10, 000, 000 primes / minute on an Intel Core i7 - 6700K CPU @ 4.00 GHz using 7 threads on 8 logical cores." +
+                 "\nUse.ReplaceTokens() to replace multiple strings in string dictionary style." +
+                 "\nUse.IsStrong() to validate strong passwords." +
+                 "\nUse.ToTimeZone() to quickly convert any given time to any given time zone." +
+                 "\nUse.Elevate() to restart current process in admin mode if it doesn't have admin rights." +
+                 "\nUse.CompoundInterest() to calculate any given interest over any given perior at any given interval." +
+                 "\nand many, many more.")]
 [assembly: AssemblyConfiguration("")]
 [assembly: AssemblyCompany("Cornelius J. van Dyk")]
-[assembly: AssemblyProduct("Extensions")]
+[assembly: AssemblyProduct("www.nuget.org/packages/Extensions.cs")]
 [assembly: AssemblyCopyright("Copyright © 2020-2021")]
 [assembly: AssemblyTrademark("Quix, Extensions.cs and Crayveon are trademarks of Crayveon Corporation.")]
 [assembly: AssemblyCulture("")]
@@ -32,7 +57,7 @@ using System.Runtime.InteropServices;
 // You can specify all the values or you can default the Build and Revision Numbers
 // by using the '*' as shown below:
 // [assembly: AssemblyVersion("1.0.1")]
-[assembly: AssemblyVersion("2.3.499")]
-[assembly: AssemblyFileVersion("2.3.499")]
+[assembly: AssemblyVersion("3.0.500")]
+[assembly: AssemblyFileVersion("3.0.500")]
 
-[assembly: AssemblyInformationalVersion("2.3.499")]
+[assembly: AssemblyInformationalVersion("3.0.500")]